requirements.txt
streamlit
pandas
Pillow
faiss-cpu
torch
transformers
gdown
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git

